fix: adjust mt_transfer* and mt_resolve to provide granular resolution
Prior to this fix, we had singular representation for token_ids that
may have many owners. There was no way to resolve approvals for those
types of tokens, typically ft style tokens.

This change links together the owners of the accounts to the ft. So
when an approval occurs we can resolve which account to transfer the
tokens from as well as handle resolution of transfer failures from
a singular token id with many owners.